Refactoring MultipleVFCreation_withMTU using new exec util:
Added a new util to run lists of commands and check exit codes in utils, cleaning up the test definitions.
Implemented unit test for this new util.
Refactored SR_IOV_MultipleVFCreation_withMTU using this new util as an example, will follow up with other tests in a new PR if this one is accepted